backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
prometheus_client==0.22.1
sqlalchemy==2.0.41
alembic==1.16.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
pydantic-settings==2.10.1
python-multipart==0.0.20
requests==2.32.4
pytest==8.4.1
scikit-learn==1.7.0
numpy==2.3.1
